change default dsn and enable upgrade from 2.8
