+++ mlrun/api/constants.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class LogSources(Enum):
+    AUTO = 'auto'
+    PERSISTENCY = 'persistency'
+    K8S = 'k8s'
