+++ mlrun/api/crud/__init__.py
@@ -1,0 +1,1 @@
+from .logs import Logs
